+++ packages/@expo/cli/src/start/server/metro/MetroBundlerDevServer.ts
@@ -1340,6 +1340,7 @@ export class MetroBundlerDevServer extends BundlerDevServer {
       await instantiateMetroAsync(this, parsedOptions, {
         isExporting: !!options.isExporting,
         exp,
+        devToolsPluginManager: this.devToolsPluginManager,
       });
 
     // Required for symbolication: